Renamed Files: RimCase to RimEclipseCase
