+++ ApplicationCode/ModelVisualization/RivCellEdgeGeometryUtils.cpp
@@ -26,7 +26,7 @@
 #include "RigGridBase.h"
 #include "RigResultAccessor.h"
 #include "RigResultAccessorFactory.h"
-#include "RimCase.h"
+#include "RimEclipseCase.h"
 #include "RimCellEdgeResultSlot.h"
 #include "RimReservoirView.h"
 #include "RimResultSlot.h"
