+++ ApplicationCode/ProjectDataModel/RimCellRangeFilterCollection.cpp
@@ -20,7 +20,7 @@
 
 #include "RimCellRangeFilterCollection.h"
 
-#include "RimCase.h"
+#include "RimEclipseCase.h"
 #include "RigGridBase.h"
 #include "RimReservoirView.h"
 #include "RigCaseData.h"
