+++ ApplicationCode/SocketInterface/RiaCaseInfoCommands.cpp
@@ -34,7 +34,7 @@
 #include "RimWellCollection.h"
 #include "Rim3dOverlayInfoConfig.h"
 #include "RimReservoirCellResultsStorage.h"
-#include "RimCase.h"
+#include "RimEclipseCase.h"
 
 #include "RigCaseData.h"
 #include "RigCaseCellResultsData.h"
